+++ lib/screens/home_navigation.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:open_source_software/managers/data_manager.dart';
 import 'package:open_source_software/managers/location_manager.dart';
 import 'package:open_source_software/managers/login_manager.dart';
 import 'package:open_source_software/screens/chatting_list.dart';
@@ -27,8 +28,7 @@ class _HomeNavigationState extends State<HomeNavigation> {
   void initState() {
     super.initState();
 
-    // 로그인 후 메인 화면에 진입한 시점에 위치 추적을 시작한다.
-    // (앱 시작 시점이 아니라 이 시점에 권한을 요청해야 사용자 맥락에 자연스럽다)
+    DataManager().initCache();
     LocationManager().initialize();
     LocationManager().onBuildingChanged = (buildingName) {
       LoginManager().updateCurrentBuilding(buildingName);